Refactor create post flow with unified CreatePostActivity
