+++ app/src/main/res/layout/create_post.xml
@@ -1,0 +1,167 @@
+<LinearLayout
+	xmlns:android="http://schemas.android.com/apk/res/android"
+	xmlns:app="http://schemas.android.com/apk/res-auto"
+	xmlns:tools="http://schemas.android.com/tools"
+	android:layout_width="match_parent"
+	android:layout_height="match_parent"
+	android:orientation="vertical"
+	android:id="@+id/_main">
+	<LinearLayout
+		android:id="@+id/main"
+		android:layout_width="match_parent"
+		android:layout_height="match_parent"
+		android:background="#FFFFFF"
+		android:orientation="vertical">
+		<LinearLayout
+			android:id="@+id/top"
+			android:layout_width="match_parent"
+			android:layout_height="wrap_content"
+			android:paddingLeft="8dp"
+			android:paddingTop="4dp"
+			android:paddingRight="8dp"
+			android:paddingBottom="4dp"
+			android:background="#FFFFFF"
+			android:gravity="center_vertical"
+			android:orientation="horizontal"
+			android:elevation="1dp">
+			<ImageView
+				android:id="@+id/back"
+				android:layout_width="48dp"
+				android:layout_height="48dp"
+				android:layout_marginRight="8dp"
+				android:padding="12dp"
+				android:src="@drawable/ic_arrow_back_black"
+				android:scaleType="fitCenter" />
+			<LinearLayout
+				android:id="@+id/topSpc"
+				android:layout_width="wrap_content"
+				android:layout_height="match_parent"
+				android:layout_marginRight="8dp"
+				android:gravity="center_vertical"
+				android:orientation="vertical"
+				android:layout_weight="1">
+				<TextView
+					android:id="@+id/title"
+					android:layout_width="wrap_content"
+					android:layout_height="wrap_content"
+					android:text="Create Post"
+					android:textSize="20sp"
+					android:textStyle="bold"
+					android:textColor="#000000" />
+			</LinearLayout>
+			<LinearLayout
+				android:id="@+id/settingsButton"
+				android:layout_width="48dp"
+				android:layout_height="48dp"
+				android:layout_marginRight="8dp"
+				android:padding="12dp"
+				android:gravity="center"
+				android:orientation="vertical">
+				<ImageView
+					android:layout_width="24dp"
+					android:layout_height="24dp"
+					android:src="@drawable/ic_settings_black"
+					android:scaleType="fitCenter" />
+			</LinearLayout>
+			<Button
+				android:id="@+id/postButton"
+				android:layout_width="wrap_content"
+				android:layout_height="wrap_content"
+				android:paddingLeft="20dp"
+				android:paddingTop="8dp"
+				android:paddingRight="20dp"
+				android:paddingBottom="8dp"
+				android:text="Post"
+				android:textSize="16sp"
+				android:textColor="@color/colorSurface"
+				style="@style/Widget.Material3.Button" />
+		</LinearLayout>
+		<LinearLayout
+			android:id="@+id/topSpace"
+			android:layout_width="match_parent"
+			android:layout_height="1dp"
+			android:padding="8dp"
+			android:background="#F5F5F5"
+			android:orientation="horizontal" />
+		<ScrollView
+			android:id="@+id/scroll"
+			android:layout_width="match_parent"
+			android:layout_height="match_parent">
+			<LinearLayout
+				android:id="@+id/scrollBody"
+				android:layout_width="match_parent"
+				android:layout_height="match_parent"
+				android:orientation="vertical">
+				<LinearLayout
+					android:id="@+id/PostInfoTop1"
+					android:layout_width="match_parent"
+					android:layout_height="wrap_content"
+					android:orientation="vertical">
+					<androidx.cardview.widget.CardView
+						android:id="@+id/imageCard"
+						android:layout_width="match_parent"
+						android:layout_height="wrap_content"
+						android:layout_margin="16dp"
+						android:gravity="center_horizontal|center_vertical"
+						app:cardElevation="2dp">
+						<LinearLayout
+							android:id="@+id/imagePlaceholder"
+							android:layout_width="match_parent"
+							android:layout_height="200dp"
+							android:gravity="center"
+							android:orientation="vertical"
+							android:background="#F5F5F5">
+							<ImageView
+								android:layout_width="48dp"
+								android:layout_height="48dp"
+								android:layout_marginBottom="8dp"
+								android:src="@drawable/ic_add_photo_black"
+								android:scaleType="fitCenter" />
+							<TextView
+								android:layout_width="wrap_content"
+								android:layout_height="wrap_content"
+								android:text="Add Photo"
+								android:textSize="16sp"
+								android:textColor="#757575" />
+						</LinearLayout>
+						<ImageView
+							android:id="@+id/postImageView"
+							android:layout_width="match_parent"
+							android:layout_height="wrap_content"
+							android:src="@drawable/default_image"
+							android:scaleType="centerCrop"
+							android:visibility="gone" />
+					</androidx.cardview.widget.CardView>
+					<com.synapse.social.studioasinc.FadeEditText
+						android:id="@+id/postDescription"
+						android:layout_width="match_parent"
+						android:layout_height="wrap_content"
+						android:layout_marginLeft="16dp"
+						android:layout_marginRight="16dp"
+						android:layout_marginBottom="16dp"
+						android:padding="16dp"
+						android:background="#F5F5F5"
+						android:textSize="18sp"
+						android:textColor="#000000"
+						android:hint="What's on your mind?"
+						android:textColorHint="#757575"
+						android:layout_weight="1"
+						android:maxLength="2500" />
+				</LinearLayout>
+				<LinearLayout
+					android:id="@+id/topSpace2"
+					android:layout_width="match_parent"
+					android:layout_height="1dp"
+					android:padding="8dp"
+					android:orientation="horizontal" />
+				<LinearLayout
+					android:id="@+id/spc2"
+					android:layout_width="match_parent"
+					android:layout_height="1dp"
+					android:padding="8dp"
+					android:background="#ECEFF1"
+					android:orientation="horizontal" />
+			</LinearLayout>
+		</ScrollView>
+	</LinearLayout>
+</LinearLayout>
